Merge voicegraph code into prealloc.
Move everything in builder/voicegraph to interp/prealloc.
Adjust data structures a little. Adapt printout code, and
in turn the player/player code using it.

Todo: The main and remaining part of the redesign planned
in 2013, expanding the code to schedule audio buffer use,
prior to audio generation, so that it can then use a flat
ordered list of what to generate and what input it needs,
instead of the old graph traversal code limited to trees.

Change log:
 * 2020-11-22: Merge voicegraph into prealloc.
 * 2020-11-24: Tidy non-NULL carrier op lists.
 * 2020-12-21: Init op lists only on (re)init,
       not for every voice or op state update.
 * 2020-12-23: Comment out the
       unused SSG_OAS_* flags.