Add SAU_getopt() and use in saugns.c.
Add and use SAU_getopt() for cli argument parsing.
The behavior is now the usual for getopt()-using programs,
and a flag and its argument can now be written without a
space to separate them. Now the use of -- to separate flags
from files is also recognized.

The getopt code is based on the public domain code by
Christopher Wellons with tweaks. Apart from style and API
changes, the arg field is always set so as to be available
for reading an optional option argument (something which is
not specified in optstring, unlike required ones).

Change common module license from ISC to 0BSD.